Fixing CPU usage bug (OpenCL code bug)

+++ resources/live3DSRRF.cl
@@ -427,7 +427,7 @@ __kernel void calculateRadialGradientConvergence(
 
 
     CGLH /= distanceWeightSum;
-    if (CGLH >= 0) CGLH = pow(CGLH, sensitivity);
+    if (CGLH >= 0) CGLH = pow((float)CGLH, (float)sensitivity);
     else CGLH = 0;
 
     // figuring out the right planes to calculate the interpolated values on
