+++ resources/liveSRRF_noInterpolation.cl
@@ -349,7 +349,7 @@ __kernel void calculateRadialGradientConvergence(
     CGLH /= distanceWeightSum; // ----------------WEIGHTING NORMALIZATION----------------
 
     // Apply Sensitivity
-    if (CGLH >= 0) CGLH = pow(CGLH, sensitivity);
+    if (CGLH >= 0) CGLH = pow((float)CGLH, (float)sensitivity);
     else CGLH = 0;
 
     float v;
